scripts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
frozenlist==1.3.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
httplib2==0.22.0
idna==3.4
loguru==0.7.0
multidict==6.0.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.5
pinecone-client==2.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
